requirements.txt
streamlit
PyMuPDF
python-docx
pytesseract
pillow
chromadb
ollama